refactor: relocate triggerToEvent function to events
