+++ src/engine/engine2/eventBus.ts
@@ -29,11 +29,9 @@ import {
   GameEvent,
   GameFightStartedEvent,
   GameTickEvent,
-  NotImplementedEvent,
   PlayerAttributeChangedEvent,
   PlayerDiedEvent,
   PlayerOverhealedEvent,
-  triggerToEventMap,
 } from "./events";
 import { getCardAttribute, getPlayerAttribute } from "./getAttribute";
 import {
@@ -281,24 +279,6 @@ export class EventBus {
   ): void {
     this.listeners.delete(eventClass);
   }
-}
-
-/**
- * Convert ability trigger to event class constructor
- */
-export function triggerToEvent(trigger: {
-  $type: TriggerType;
-}): GameEventConstructor<GameEvent> {
-  const triggerType = trigger?.$type || "";
-  const eventClass = triggerToEventMap[triggerType];
-
-  if (!eventClass) {
-    console.warn(`Unhandled trigger type: ${triggerType}`);
-    throw new Error(`Unhandled trigger type: ${triggerType}`);
-    return NotImplementedEvent;
-  }
-
-  return eventClass;
 }
 
 function checkSubjectAndTargetIfDefined(
